# be-lemdiklat-permapendis/Dto/CreateUserRequest.cs
namespace be_lemdiklat_permapendis.Dto;

public class CreateUserRequest
{
    public string Username { get; set; } = string.Empty;
    public string Password { get; set; } = string.Empty;
    public string Name { get; set; } = string.Empty;
    public string? Email { get; set; }
    public string? Phone { get; set; }
    public string? Address { get; set; }
    public string? City { get; set; }
    public int RoleId { get; set; }
}